# Transition and memoize course cards