@firejake308 2021-02-08
Merged with master. The transition between basic and section select is going to be trickier than I first thought because we're currently using Material UI's `Collapse` element to manage the transition for us, which only has 2 possible states: collapsed and expanded, and transitions only occur when that state changes. Since Basic Select and Section Select are both included in the expanded state, there is no transition between them. If it is doable, it will need to be a separate PR.