@firejake308 2021-02-17
@gannonprudhomme I figured out what's causing the bug that you pointed out with the last card's unnecessary collapse animation:

When the page is first created, the CourseSelectColumn has a single, empty card that is expanded by default. On the user's first visit, this works nicely because now the card is big and they can see where to enter the course. However, if the user already has saved courses, then the collapsed state of the saved courses overwrites the expanded state of the formerly empty 0th card. Thus, the 0th card has gone from expanded to collpased, so the collapse animation plays.

The ideal behavior would be for the 0th card to start out collapsed if the user had saved courses and expanded if the user did not have any saved courses. However, it is not possible for us to know that until we make the `/sessions/get_saved_courses` API call, so we need to decide on the best initial state to use until then. If we change the 0th card to be collapsed by default, we run the risk of confusing first-time users if `/sessions/get_saved_courses` takes too long to return. However, if we change the 0th card to be expanded by default, we have to play an extra animation to collapse the 0th card and expand whatever card the user was looking at when they left.

What's y'all's opinion? I'm thinking the initial API call shouldn't take too long, so making the blank card collpased until it returns shouldn't be too much of an issue. However, I'd like y'all's input if y'all think leaving it as is would be better or if there's a better approach that I haven't thought of.