@firejake308 2021-03-01
@gannonprudhomme I updated an outdated CSS rule that seems to improve transition smoothness on my computer. Can you confirm that it looks better for you too?