@gannonprudhomme 2021-03-01
Yep that seems to have fixed it! Heck yeah