@firejake308 2021-03-25
Fixed the bug that you pointed out. However, there is still a bug when you have a large course card open and then add a new card. In that case, the card's contents are rendered like 8px above where they should be.