@firejake308 2021-04-15
Ironed out *most* of the bugs. The opening animation does feel a little fast on my computer, and it does a noticeable adjsutment when it realizes that it needs a scroll bar, but I think these are acceptable caveats. Unless y'all have any idea what might be making the opening transition too fast?