@firejake308 2021-04-22
@gannonprudhomme Fixed the speed by adding a `React.memo`