@firejake308 2021-04-27
> I found one bug in this, which is that whenever you close a card, the next time you expand one it won't play the animation. This happens on the next time a card is expanded whether there was one expanded when you closed a card or not.

I swear this PR is the hydra of bugs. Every time I squash one bug, two more appear in its place. That is definitely something that needs fixing, but in the interest of launch, I'll make an issue for it and fix it later